requirements.txt
# ============================================================
# Drowning Detection System - Requirements
# ระบบตรวจจับการจมน้ำ - YOLOv11 Standard + ID Tracking
# ============================================================

# --- Core AI & Computer Vision (YOLO11) ---
ultralytics>=8.3.0
torch>=2.4.0
torchvision>=0.19.0
opencv-python>=4.10.0
numpy>=1.26.4

# --- Configuration & Environment ---
python-dotenv>=1.0.1

# --- Telegram Bot (Async) ---
python-telegram-bot>=21.4